% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{.array_to_sd_image}
\alias{.array_to_sd_image}
\title{Convert R array [H, W, 3] to sd_image list}
\usage{
.array_to_sd_image(arr)
}
\arguments{
\item{arr}{3D numeric array [height, width, channels] in [0, 1]}
}
\value{
SD image list (width, height, channel, data)
}
\description{
Convert R array [H, W, 3] to sd_image list
}
\keyword{internal}
