% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{.blend_mask}
\alias{.blend_mask}
\title{Build linear blend mask for a patch}
\usage{
.blend_mask(h, w, overlap, is_left, is_top, is_right, is_bottom)
}
\arguments{
\item{h}{Patch height}

\item{w}{Patch width}

\item{overlap}{Overlap in pixels}

\item{is_left, is_top, is_right, is_bottom}{Whether patch is at canvas edge}
}
\value{
Matrix [h, w] with blend weights in [0, 1]
}
\description{
Build linear blend mask for a patch
}
\keyword{internal}
