% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_manager.R
\name{.guess_component}
\alias{.guess_component}
\title{Guess component role from filename}
\usage{
.guess_component(filename)
}
\arguments{
\item{filename}{File basename}
}
\value{
Character: "diffusion", "vae", "clip_l", "clip_g", "t5xxl",
  "taesd", or "unknown"
}
\description{
Guess component role from filename
}
\keyword{internal}
