% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_manager.R
\name{.guess_model_type}
\alias{.guess_model_type}
\title{Guess model type from filename}
\usage{
.guess_model_type(filename)
}
\arguments{
\item{filename}{File basename}
}
\value{
Character: "flux2", "flux", "sdxl", "sd1", "sd2", "sd3", or "unknown"
}
\description{
Guess model type from filename
}
\keyword{internal}
