% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{.images_to_base64}
\alias{.images_to_base64}
\title{Encode sd_image list to base64 PNG strings}
\usage{
.images_to_base64(images)
}
\arguments{
\item{images}{List of sd_image objects}
}
\value{
Character vector of base64-encoded PNG strings
}
\description{
Encode sd_image list to base64 PNG strings
}
\keyword{internal}
