% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{.native_tile_size}
\alias{.native_tile_size}
\title{Get native tile size for a model type}
\usage{
.native_tile_size(model_type)
}
\arguments{
\item{model_type}{One of "sd1", "sd2", "sdxl", "flux", "flux2", "sd3"}
}
\value{
Integer tile size in pixels
}
\description{
Get native tile size for a model type
}
\keyword{internal}
