% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{.resize_sd_image}
\alias{.resize_sd_image}
\title{Bilinear resize of an SD image}
\usage{
.resize_sd_image(image, target_w, target_h)
}
\arguments{
\item{image}{SD image list}

\item{target_w}{Target width}

\item{target_h}{Target height}
}
\value{
Resized SD image
}
\description{
Bilinear resize of an SD image
}
\keyword{internal}
