% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{.unbox_scalars}
\alias{.unbox_scalars}
\title{Recursively unbox scalar values in nested lists for JSON serialization}
\usage{
.unbox_scalars(x, keep_arrays = character(0))
}
\arguments{
\item{x}{List or atomic value}
}
\value{
Same structure with scalars wrapped in \code{jsonlite::unbox}
}
\description{
Recursively unbox scalar values in nested lists for JSON serialization
}
\keyword{internal}
