% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{LORA_APPLY_MODE}
\alias{LORA_APPLY_MODE}
\title{LoRA apply modes}
\format{
An object of class \code{list} of length 3.
}
\usage{
LORA_APPLY_MODE
}
\description{
LoRA apply modes
}
\keyword{datasets}
