% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_manager.R
\name{sd_list_models}
\alias{sd_list_models}
\title{List registered models}
\usage{
sd_list_models()
}
\value{
Data frame with columns: id, model_type, loaded, diffusion_path
}
\description{
Returns a data frame of all models in \code{~/.sd2R/models.json},
with a column indicating which are currently loaded in memory.
}
