% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_utils.R
\name{sd_load_image}
\alias{sd_load_image}
\title{Load image from file as SD image}
\usage{
sd_load_image(path, channels = 3L)
}
\arguments{
\item{path}{Path to image file (PNG)}

\item{channels}{Number of output channels (3 for RGB, default)}
}
\value{
SD image list (width, height, channel, data as raw vector)
}
\description{
Reads a PNG file and converts it to the SD image format
(list with width, height, channel, data) suitable for img2img.
}
