% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{sd_load_pipeline}
\alias{sd_load_pipeline}
\title{Load pipeline from JSON}
\usage{
sd_load_pipeline(path)
}
\arguments{
\item{path}{Path to a JSON file saved by \code{\link{sd_save_pipeline}}.}
}
\value{
An \code{sd_pipeline} object.
}
\description{
Load pipeline from JSON
}
