% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{sd_node}
\alias{sd_node}
\title{Create a pipeline node}
\usage{
sd_node(type, ...)
}
\arguments{
\item{type}{Node type: \code{"txt2img"}, \code{"img2img"}, \code{"upscale"},
or \code{"save"}.}

\item{...}{Parameters for the node (passed to the corresponding function).}
}
\value{
A list with class \code{"sd_node"}.
}
\description{
Create a pipeline node
}
