% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{sd_pipeline}
\alias{sd_pipeline}
\title{Create a pipeline from nodes}
\usage{
sd_pipeline(...)
}
\arguments{
\item{...}{\code{sd_node} objects in execution order.}
}
\value{
A list with class \code{"sd_pipeline"}.
}
\description{
Nodes are executed sequentially. The image output of each node is passed
as input to the next node.
}
