% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{sd_preview_stop}
\alias{sd_preview_stop}
\title{Disable live generation previews}
\usage{
sd_preview_stop()
}
\value{
Invisibly \code{NULL}.
}
\description{
Removes the preview callback and cleans up the temporary \code{.tmp} file.
}
\seealso{
\code{\link{sd_preview_start}}
}
