% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_utils.R
\name{sd_save_image}
\alias{sd_save_image}
\title{Save SD image to PNG file}
\usage{
sd_save_image(image, path)
}
\arguments{
\item{image}{SD image (list with width, height, channel, data) as returned
by \code{sd_txt2img()} or \code{sd_img2img()}. Can also be a 3D numeric
array [height, width, channels] with values in [0, 1].}

\item{path}{Output file path (should end in .png)}
}
\description{
Save SD image to PNG file
}
