% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{sd_save_pipeline}
\alias{sd_save_pipeline}
\title{Save pipeline to JSON}
\usage{
sd_save_pipeline(pipeline, path)
}
\arguments{
\item{pipeline}{An \code{sd_pipeline} object.}

\item{path}{File path (should end in \code{.json}).}
}
\description{
Save pipeline to JSON
}
