% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{SD_TYPE}
\alias{SD_TYPE}
\title{Weight types (ggml quantization types)}
\format{
An object of class \code{list} of length 35.
}
\usage{
SD_TYPE
}
\description{
Weight types (ggml quantization types)
}
\keyword{datasets}
