% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{sd_upscale_image}
\alias{sd_upscale_image}
\title{Upscale an image using ESRGAN}
\usage{
sd_upscale_image(esrgan_path, image, upscale_factor = 4L, n_threads = 0L)
}
\arguments{
\item{esrgan_path}{Path to ESRGAN model file}

\item{image}{SD image to upscale (list with width, height, channel, data)}

\item{upscale_factor}{Upscale factor (default 4)}

\item{n_threads}{Number of CPU threads (0 = auto-detect)}
}
\value{
Upscaled SD image
}
\description{
Upscale an image using ESRGAN
}
