% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sd_vulkan_device_count}
\alias{sd_vulkan_device_count}
\title{Get number of Vulkan GPU devices}
\usage{
sd_vulkan_device_count()
}
\value{
Integer, number of Vulkan devices (0 if Vulkan is not available)
}
\description{
Returns the number of Vulkan-capable GPU devices available on the system.
Useful for deciding whether to use \code{\link{sd_generate_multi_gpu}}.
}
